+++ src/main/frontend/src/assets/main.scss
@@ -14,6 +14,8 @@ $dropdown-link-hover-bg: $primary;
 
 .btn-link {
   text-decoration: none;
+  padding-left: 0;
+  padding-right: 0;
 }
 
 .btn-link:hover {
