+++ src/main/frontend/src/components/FooterComponent.vue
@@ -1,0 +1,36 @@
+<!--
+  - Copyright 2023 Red Hat, Inc.
+  -
+  - Red Hat licenses this file to you under the Apache License, version 2.0
+  - (the "License"); you may not use this file except in compliance with the
+  - License.  You may obtain a copy of the License at:
+  -
+  - http://www.apache.org/licenses/LICENSE-2.0
+  -
+  - Unless required by applicable law or agreed to in writing, software
+  - distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  - WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+  - License for the specific language governing permissions and limitations
+  - under the License.
+  -->
+
+<template>
+  <div class="row mt-5">
+    <div class="col-sm-12">
+      <p class="small text-end">
+        <a href="https://how-to.vertx.io/"
+          ><i class="bi-arrow-right" aria-hidden="true"></i> Find a Vert.x how-to</a
+        >
+      </p>
+    </div>
+  </div>
+  <div class="row">
+    <div class="col-sm-12">
+      <p class="small text-end">
+        <a href="https://github.com/vert-x3/vertx-starter/issues"
+          ><i class="bi-bug"></i> Report an issue</a
+        >
+      </p>
+    </div>
+  </div>
+</template>
